refactoring converter tests
