+++ spring-cloud-gcp-data-spanner/src/main/java/org/springframework/cloud/gcp/data/spanner/core/convert/SpannerEntityWriter.java
@@ -1,0 +1,40 @@
+/*
+ *  Copyright 2018 original author or authors.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+package org.springframework.cloud.gcp.data.spanner.core.convert;
+
+import java.util.Set;
+
+import com.google.cloud.spanner.Mutation;
+import com.google.cloud.spanner.Mutation.WriteBuilder;
+
+import org.springframework.data.convert.EntityWriter;
+
+/**
+ * @author Balint Pato
+ */
+public interface SpannerEntityWriter extends EntityWriter<Object, Mutation.WriteBuilder> {
+
+	/**
+	 * Writes an object's properties to the sink.
+	 * @param source the object to write
+	 * @param sink the sink to which to write
+	 * @param includeColumns the properties/columns to write. If null, then all columns are
+	 * written.
+	 */
+	void write(Object source, WriteBuilder sink, Set<String> includeColumns);
+
+}
